refactor(site): split spend page sections
